add store object info, improve om top-level ux
- om.stats(), store.stats() provide storage object information
- improve om top-level module ux: om.buckets(), om.list(), om.stats(),
om.metadata(), om.help()
- mixins now provide the base cls name instead of the top-level
(improves help())

+++ omegaml/store/streams.py
@@ -103,12 +103,13 @@ class StreamsProxy(OmegaStore):
     def getl(self, name, **kwargs):
         return self.get(name, lazy=True, **kwargs)
 
-    def drop(self, name, force=False):
+    def drop(self, name, force=False, keep_data=True, **kwargs):
         meta = self.metadata(name)
         if meta is None and not force:
             raise DoesNotExist()
         stream = self.get(name)
-        stream.buffer().delete()
+        if not keep_data:
+            stream.buffer().delete()
         stream.delete()
         meta.delete() if meta is not None else None
         return True
